Grammar: Use classes instead of assoc arrays to reduce memory use
In the tokenizer cache, replace associative arrays with a value class.
It turns out that an instance of a class with 4 properties takes less
memory than an associative array with 2 items, at least on PHP 7.4.

As a result, memory use of the parser (as a whole) is lower by up to
35% on my machine, in this denegerate test case: (see T353874)

	php -r 'echo str_repeat("!", 100_000);' | php bin/parse.php >/dev/null

WikiPEG is bumped to 4.0.0

Originally prototyped in I213fe4f2cedd579a4aa55efdb783f1968400ab9a

Depends-On: I691fadecd17f28923de429f0a0f41605a76fe60a
Change-Id: I7c0931e70d3f037c357996b8b3c31dc75eb038f4