Split first argument from target in parser function ParamInfo
The ParamInfo objects generated for parser function invocations used
to have the first argument glued to the target in the
TemplateInfo::$targetWt property.  Now split off the first argument
and create a proper ParamInfo for it, and record the *actual*
target in TemplateInfo::$targetWt.

This only affects our internal representation; for
backward-compatibility we glue the pieces together when serializing
TemplateInfo/ParamInfo, but this makes the internal representation
easier to work with.

This patch also changes TemplateInfo::$paramInfos from a map
to an ordered list, which allows us to properly represent
transclusions with duplicated argument keys.  We also add
'order' and 'eq' properties for parser functions (T390344) as
documented in

  https://www.mediawiki.org/wiki/Parsoid/MediaWiki_DOM_spec/Parser_Functions

to ensure that parameter order and named argument status is preserved.
Preserving this information for other transclusion types is
done via a mechanism involving `data-parsoid.pi` (T404772)
and so since we're adding a "new" first argument for parser
functions a number of parser tests get data-parsoid additions like
`"pi":[{"k":"1"}]` in their known differences output.

In addition, there is one parser test which adds <nowiki>
around a problematic `]]` sequence now that the first
argument is actually serialized as an argument (and not
as part of the target).

Bug: T390344
Depends-On: I506b666ffa7ae280e69bb567c59cbb16366e650b
Change-Id: I1c78a6d67e63133bdd338d64e62ca2b19374762a